+++ src/components/custom-field-input-multiple-choice/custom-field-input-multiple-choice.jsx
@@ -9,11 +9,20 @@ import FormControl from '../form-control/form-control.jsx';
 import Icon from '../icon/icon.jsx';
 import iconCaretDown from '../../svgs/icon-caret-down.svg';
 import iconCaretDownDisabled from '../../svgs/icon-caret-down-disabled.svg';
+import iconCaution from '../../svgs/icon-caution-fill.svg';
+import iconClear from '../../svgs/icon-clear-small.svg';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
 import TagList from '../tag-list/tag-list.jsx';
 import Tag from '../tag/tag.jsx';
 import styles from './custom-field-input-multiple-choice.css';
+
+function getClassName(readOnly, helpText) {
+  if (readOnly) return styles['read-only-container'];
+  if (helpText) return styles['invalid-container'];
+
+  return styles['read-write-container'];
+}
 
 function CustomFieldInputMultipleChoice(props) {
   const autocompleteRef = useRef();
@@ -25,9 +34,7 @@ function CustomFieldInputMultipleChoice(props) {
     .filter(choice => !value.includes(choice));
   const choicesRefs = visibleChoices.map(() => createRef());
   const valueRefs = value.map(() => createRef());
-  const classContainer = props.readOnly ?
-    styles['read-only-container'] :
-    styles['read-write-container'];
+  const classContainer = getClassName(props.readOnly, props.helpText);
   const renderPopup = !props.readOnly && expanded && visibleChoices.length !== 0;
 
   function onChoiceRemove(event) {
@@ -45,12 +52,20 @@ function CustomFieldInputMultipleChoice(props) {
     setAutocompleteValue('');
   }
 
+  function onChoicesClear(event) {
+    event.preventDefault();
+    setValue([]);
+    setExpanded(false);
+  }
+
   function onAutocompleteChange(event) {
     setExpanded(true);
     setAutocompleteValue(event.target.value);
   }
 
-  function onClick() {
+  function onClick(event) {
+    if (event.defaultPrevented) return;
+
     setExpanded(true);
   }
 
@@ -71,43 +86,73 @@ function CustomFieldInputMultipleChoice(props) {
 
   return (
     <FormControl
+      error={props.helpText}
       label={props.label}
       labelId={`${props.id}-label`}
       id={`${props.id}-autocomple`}
       onKeyDown={onKeyDown}
       readOnly={props.readOnly}
     >
-      <TagList
+      <div
         className={classContainer}
-        id={props.id}
-        labelledBy={`${props.id}-label`}
         onClick={onClick}
-        refs={valueRefs}
+        role="presentation"
       >
-        {value.map((choice, index) => (
-          <Tag
-            defaultActive={index === 0}
-            id={`${props.id}-${choice.id}`}
-            key={`${props.id}-${choice.id}`}
-            onRemove={onChoiceRemove}
-            readOnly={props.readOnly}
-            ref={valueRefs[index]}
-          >
-            {choice.label}
-          </Tag>
-        ))}
-        {!props.readOnly && (
-          <input
-            aria-labelledby={`${props.id}-label`}
-            className={styles['autocomplete-input']}
-            id={`${props.id}-autocomple`}
-            onChange={onAutocompleteChange}
-            ref={autocompleteRef}
-            value={autocompleteValue}
+        <TagList
+          className={styles['tag-list']}
+          id={props.id}
+          labelledBy={`${props.id}-label`}
+          refs={valueRefs}
+        >
+          {value.map((choice, index) => (
+            <Tag
+              defaultActive={index === 0}
+              id={`${props.id}-${choice.id}`}
+              key={`${props.id}-${choice.id}`}
+              onRemove={onChoiceRemove}
+              readOnly={props.readOnly}
+              ref={valueRefs[index]}
+            >
+              {choice.label}
+            </Tag>
+          ))}
+          {!props.readOnly && (
+            <input
+              aria-labelledby={`${props.id}-label`}
+              className={styles['autocomplete-input']}
+              id={`${props.id}-autocomple`}
+              onChange={onAutocompleteChange}
+              ref={autocompleteRef}
+              value={autocompleteValue}
+            />
+          )}
+        </TagList>
+        <div className={styles['icons-container']}>
+          {!props.readOnly && props.helpText && (
+            <Icon
+              className={styles.icon}
+              currentColor="caution"
+              fill="skip"
+              name={iconCaution.id}
+            />
+          )}
+          {!props.readOnly && value.length > 0 && (
+            <Icon
+              className={styles['clear-icon']}
+              fill="skip"
+              name={iconClear.id}
+              onClick={onChoicesClear}
+              ariaLabel={`Remove all selected choices on ${props.label}`}
+              role="button"
+            />
+          )}
+          <Icon
+            className={styles.icon}
+            name={props.readOnly ? iconCaretDownDisabled.id : iconCaretDown.id}
+            fill="skip"
           />
-        )}
-        <Icon className={styles['input-icon']} name={props.readOnly ? iconCaretDownDisabled.id : iconCaretDown.id} fill="skip" />
-      </TagList>
+        </div>
+      </div>
       {(renderPopup &&
         <Listbox
           className={styles['popup-container']}
@@ -135,6 +180,7 @@ CustomFieldInputMultipleChoice.propTypes = {
     id: PropTypes.string.isRequired,
     label: PropTypes.string.isRequired,
   })).isRequired,
+  helpText: PropTypes.string,
   id: PropTypes.string.isRequired,
   label: PropTypes.string.isRequired,
   readOnly: PropTypes.bool,
@@ -145,6 +191,7 @@ CustomFieldInputMultipleChoice.propTypes = {
 };
 
 CustomFieldInputMultipleChoice.defaultProps = {
+  helpText: undefined,
   readOnly: false,
   value: [],
 };