+++ src/components/icon/icon.test.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import Icon from './icon.jsx';
 
 describe('Icon', () => {
@@ -39,6 +40,15 @@ describe('Icon', () => {
         <Icon {...requiredProps} id="unique-id" />
       ));
       expect(screen.getByRole('img')).toHaveAttribute('id', 'unique-id');
+    });
+  });
+
+  describe('onClick API', () => {
+    it('can be set', () => {
+      const onClickSpy = jest.fn();
+      render(<Icon {...requiredProps} onClick={onClickSpy} />);
+      userEvent.click(screen.getByRole('img'));
+      expect(onClickSpy).toHaveBeenCalledWith(expect.anything());
     });
   });